requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers
python-dotenv
langchain-text-splitters